+++ R/satscan-configuration.R
@@ -41,7 +41,7 @@
 #'
 #' ## Wrangle data with `{mwana}` ----
 #' x <- anthro |>
-#'   dplyr::rename(longitude = x, latitude = y) |>
+#'   dplyr::rename(longitude = y, latitude = x) |>
 #'   mwana::mw_wrangle_wfhz(
 #'     sex = sex,
 #'     .recode_sex = TRUE,
@@ -59,7 +59,9 @@
 #'   .data = x,
 #'   filename = "Locality",
 #'   dir = directory,
-#'   .gam_based = "wfhz"
+#'   .gam_based = "wfhz",
+#'   latitude = latitude,
+#'   longitude = longitude
 #' )
 #'
 #' library(rsatscan) # important to make `{wowi}` access `{rsatscan}`-specific eviroment