+++ tests/testthat/test-data-wranglers.R
@@ -13,7 +13,7 @@ testthat::test_that(
   {
     ## Sample data ----
     x <- anthro |>
-      dplyr::rename(longitude = x, latitude = y) |>
+      dplyr::rename(longitude = y, latitude = x) |>
       dplyr::filter(district == "Kotido") |>
       mwana::mw_wrangle_wfhz(
         sex = sex,
@@ -28,7 +28,12 @@ testthat::test_that(
       )
 
     ## Observed results ----
-    r <- do.call(what = wrangle_data, args = list(.data = x, .gam_based = "wfhz"))
+    r <- wrangle_data(
+      .data = x,
+      latitude = latitude,
+      longitude = longitude,
+      .gam_based = "wfhz"
+    )
 
     ## Test checks ----
     testthat::expect_type(object = r, type = "list")
@@ -52,7 +57,7 @@ testthat::test_that(
   {
     ## Sample data ----
     x <- anthro |>
-      dplyr::rename(longitude = x, latitude = y) |>
+      dplyr::rename(longitude = y, latitude = x) |>
       dplyr::filter(district == "Kotido") |>
       mwana::mw_wrangle_age(age = age) |>
       mwana::mw_wrangle_muac(
@@ -68,9 +73,11 @@ testthat::test_that(
 
 
     ## Observed results ----
-    r <- do.call(
-      what = wrangle_data,
-      args = list(.data = x, .gam_based = "muac")
+    r <- wrangle_data(
+      .data = x,
+      latitude = latitude,
+      longitude = longitude,
+      .gam_based = "muac"
     )
 
     ## Test checks ----
@@ -95,7 +102,7 @@ testthat::test_that(
   {
     ### Observed data ----
     x <- anthro |>
-      dplyr::rename(longitude = x, latitude = y) |>
+      dplyr::rename(longitude = y, latitude = x) |>
       dplyr::filter(district == "Kotido") |>
       mwana::mw_wrangle_wfhz(
         sex = sex,
@@ -121,9 +128,11 @@ testthat::test_that(
       )
 
     ### Observed results ----
-    r <- do.call(
-      what = wrangle_data,
-      args = list(.data = x, .gam_based = "combined")
+    r <- wrangle_data(
+      .data = x,
+      latitude = latitude,
+      longitude = longitude,
+      .gam_based = "combined"
     )
 
     ### Test checks ----
@@ -152,7 +161,7 @@ testthat::test_that(
   {
     ### Sample data ----
     x <- anthro |>
-      dplyr::rename(longitude = x, latitude = y) |>
+      dplyr::rename(longitude = y, latitude = x) |>
       dplyr::filter(district == "Kotido") |>
       mwana::mw_wrangle_wfhz(
         sex = sex,
@@ -171,14 +180,13 @@ testthat::test_that(
     out_dir <- file.path(tmp, "input-files") # this will be the dir
 
     ### Observed results ----
-    do.call(
-      what = ww_wrangle_data,
-      args = list(
-        .data = x,
-        filename = "localityA",
-        dir = out_dir,
-        .gam_based = "wfhz"
-      )
+    ww_wrangle_data(
+      .data = x,
+      filename = "localityA",
+      dir = out_dir,
+      .gam_based = "wfhz",
+      latitude = latitude,
+      longitude = longitude
     )
 
     ## The tests ----
@@ -186,15 +194,15 @@ testthat::test_that(
     testthat::expect_true(file.exists(file.path(out_dir, "localityA.ctl")))
     testthat::expect_true(file.exists(file.path(out_dir, "localityA.geo")))
     testthat::expect_message(
-      object = do.call(
-        what = ww_wrangle_data,
-        args = list(
-          .data = x,
-          filename = "localityA",
-          dir = out_dir,
-          .gam_based = "wfhz"
-        )
-      ), regexp = paste0("`", basename(out_dir), "` already exists in project repo."), fixed = TRUE
+      ww_wrangle_data(
+        .data = x,
+        filename = "localityA",
+        dir = out_dir,
+        latitude = latitude,
+        longitude = longitude,
+        .gam_based = "wfhz"
+      ),
+      regexp = paste0("`", basename(out_dir), "` already exists in project repo."), fixed = TRUE
     )
   }
 )