+++ NEWS.md
@@ -1,3 +1,13 @@
+# wowi 1.0.0
+
+## New features ✨
+
+* Added a Shiny application.
+
+## Bug fixes 
+
+* Resolved mislabelled geographic coordinates.
+
 # wowi 0.1.0
 
 * Initial pre-release version for alpha-testing.